Fix class name

+++ program/skeleton/src/Entity/EV3/Devise/InputDevise/Sensor/JyroSensor.java
@@ -1,15 +1,19 @@
 package Entity.EV3.Devise.InputDevise.Sensor;
 
-public class JyroSensor extends Sensor {
-  public JyroSensor(char port) {
+public class JyroSensor extends Sensor
+{
+  public JyroSensor(char port)
+  {
     super(port);
   }
 
-  public int getValue() {
+  public int getValue()
+  {
     return 0;
   }
 
-  public float getFloatValue() {
-    return (float)0.0;
+  public float getFloatValue()
+  {
+    return 0.0F;
   }
 }
